+++ source/CoordinateConversion/ArcMapAddinCoordinateConversion/ViewModels/ConvertTabViewModel.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using CoordinateConversionLibrary.Helpers;
+using CoordinateConversionLibrary.Models;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -28,14 +29,24 @@ namespace ArcMapAddinCoordinateConversion.ViewModels
         {
             InputCoordinateHistoryList = new ObservableCollection<string>();
 
+            // commands
+            AddNewOCCommand = new RelayCommand(OnAddNewOCCommand);
             CopyAllCommand = new RelayCommand(OnCopyAllCommand);
         }
 
         public ObservableCollection<string> InputCoordinateHistoryList { get; set; }
 
+        public RelayCommand AddNewOCCommand { get; set; }
         public RelayCommand CopyAllCommand { get; set; }
 
-        internal void OnCopyAllCommand(object obj)
+        private void OnAddNewOCCommand(object obj)
+        {
+            // Get name from user
+            string name = CoordinateType.DD.ToString();
+            Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.AddNewOutputCoordinate, new OutputCoordinateModel() { Name = name, CType = CoordinateType.DD, Format = "Y0.0#N X0.0#E" });
+        }
+        
+        private void OnCopyAllCommand(object obj)
         {
             Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.CopyAllCoordinateOutputs, InputCoordinate);
         }
