+++ source/CoordinateConversion/CoordinateConversionLibrary/Views/CollectTabView.xaml
