+++ source/CoordinateConversion/CoordinateConversionLibrary/Views/CollectTabView.xaml.cs
